разнес логику по файлам и убрал колхоз
