+++ internal/storage/repository.go
@@ -10,7 +10,3 @@ type UserRepository interface {
 	CreateUser(ctx context.Context, user domain.User) (int, error)
 	GetUserByID(ctx context.Context, id int) (domain.User, error)
 }
-
-// интерфейс для Product
-
-// интерфейс для Order
